+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - CLI: remove the scopes column from the `list-users` command
 - CLI: the `username` parameter is now required for the `create-user` command
+- CLI: add the `read-user` command
 
 ### Fixed