+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-05-31
+
 ### Added
 
 - Track user last login date time
@@ -101,7 +103,8 @@ and this project adheres to
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.7.0...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.8.0...main
+[0.8.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.4.0...v0.5.0